pyproject.toml
[project]
name = "hotelbackend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.2.1",
    "datetime>=5.5",
    "fastapi>=0.115.6",
    "passlib>=1.7.4",
    "pip>=24.3.1",
    "pydantic[email]>=2.10.5",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]
